% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hpsi_data.R
\name{get_hpsi_data}
\alias{get_hpsi_data}
\title{Get home purchase sentiment index data}
\usage{
get_hpsi_data(age_group = NA)
}
\arguments{
\item{age_group}{a number between 1 and 4 which may be passed into the function
as a character. By default, \code{age_group} is NULL, which specifies all ages aggregated
into one group. Specifying an age group of 1 refers to the 18-34 years old, 2 refers
to 35-44, 3 refers to 45-64, and 4 refers to 65 years old and older.}
}
\value{
A \code{tbl_df} with the columns \code{date}, \code{age_group}, and \code{hpsi_value}.
}
\description{
Get home purchase sentiment index data
}
\details{
The home purchase sentiment index (HPSI) is an index calculated
using Fannie Mae's National Housing Survey data to represent consumer's
home purchase sentiment. Details can be found
\href{http://www.fanniemae.com/resources/file/research/housingsurvey/pdf/hpsi-overview.pdf}{here}.
}
\examples{
## This gets HPSI data for all ages aggregated
hpsi_all_ages <- get_hpsi_data()

## getting HPSI data for respondents aged 18-34 years old
hpsi_18_34 <- get_hpsi_data(age_group = 1)

## query the API 4 times to get data for all age groups unaggregated
hpsi_all_unagg <- purrr::map_dfr(1:4, get_hpsi_data)

}
\seealso{
\link{get_nhs_data}
}
