% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator.R
\name{get_indicator_all}
\alias{get_indicator_all}
\alias{get_indicator}
\title{Get economic or housing indicators}
\usage{
get_indicator_all(category = c("economic", "housing"), indicator)

get_indicator(category = c("economic", "housing"), indicator)
}
\arguments{
\item{category}{Either "economic" or "housing"}

\item{indicator}{One of the indicator series available under housing or economic indicators. See 'Details' for options.}
}
\value{
A \code{grouped_df} that is grouped by the column \code{report_date}.
}
\description{
Get economic or housing indicators
}
\details{
This function has access to the following indicator series for \strong{economic} specific data:
\itemize{
\item gross-domestic-product
\item personal-consumption-expenditures
\item residential-fixed-investment
\item business-fixed-investment
\item government-consumption-and-investment
\item net-exports
\item change-in-business-inventories
\item gdp-price-index
\item consumer-price-index
\item consumer-price-index-excl-food-and-energy
\item personal-chain-expenditures-chain-price-index
\item personal-chain-expenditures-chain-price-index-excl-food-and-energy
\item unemployment-rate-percent
\item employment-total-nonfarm-percent-change-quarterly-saar-annual-q4-q4
\item federal-funds-rate
\item 1-year-treasury-note-yield
\item 10-year-treasury-bond-yield
}

Data specific to \strong{housing} includes:
\itemize{
\item fannie-mae-hpi
\item federal-housing-finance-agency-purchase-only-index
\item 30-year-fixed-rate-mortgage
\item 5-year-adjustable-rate-mortgage
\item 1-year-adjustable-rate-mortgage
\item mortgage-originations-combined
\item mortgage-originations-purchase
\item mortgage-originations-refinance
\item mortgage-originations-refinance-share-percent
\item housing-starts-all
\item housing-starts-single-family
\item housing-starts-multi-family
\item new-single-family-home-sales
\item total-existing-home-sales
\item total-home-sales
\item median-new-home-price
\item median-total-existing
}

The difference between \code{get_indicator} and \code{get_indicator_all} is that the latter gets
data for all report dates. For example, \code{get_indicator_all} will include revised forecasts over time.
The former will get you the latest report with end of year values removed.
}
\examples{
get_indicator("economic", "gross-domestic-product")
get_indicator("housing", "housing-starts-single-family")

## This gets all reports on GDP and single family housing starts
get_indicator_all("economic", "gross-domestic-product")
get_indicator_all('housing', 'housing-starts-single-family')

## quick plotting is better suited with `get_indicator`
gdp <- get_indicator('economic', 'gross-domestic-product')
ggplot(gdp,
  aes(x = indicator_val_dt, y = value, linetype = forecast)
) +
geom_line(size = 1)

}
\seealso{
\code{\link[=get_url]{get_url()}}
}
