% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lender_sentiment.R
\name{get_lender_sentiment}
\alias{get_lender_sentiment}
\title{Get Mortgage Lender Sentiment Survey data}
\usage{
get_lender_sentiment()
}
\value{
A \code{tbl_df} containing the following columns
\itemize{
\item \code{question} -- The question number on the survey
\item \code{lender_catg} -- A category defining the lender group.
\item \code{survey_date} -- The date of the survey
\item \code{percent_net_up} -- This is equivalent to \code{percent_went_up} minus \code{percent_went_down}
\item \code{percent_stays_the_same} -- The proportion of lenders that answered the question as
"stays the same."
\item \code{percent_went_down} -- The proportion of lenders that answered the question as "went down."
\item \code{percent_went_up} -- The proportion of lenders that answered the question as "went up."
}
}
\description{
Get Mortgage Lender Sentiment Survey data
}
\details{
The questionnaire for the mortgage lender sentiment survey is available on the
\href{http://www.fanniemae.com/resources/file/research/mlss/pdf/mortgage-lender-sentiment-survey-questionnaire-q42018.pdf}{Fannie Mae website}.
Available questions in the survey data are:
\itemize{
\item 6 -- Over the \strong{past three months}, apart from normal seasonal variation,
did your firm’s consumer demand for single-family \strong{purchase} mortgages
go up, go down, or stay the same?
\itemize{
\item Part a refers to GSE Eligible mortgages
\item Part b refers to Non-GSE Eligible mortgages
\item Part c refers to government (i.e. Federal Housing Administration
or Department of Veteran Affairs) mortgages.
}
\item 10 -- Over the \strong{past three months}, apart from normal seasonal variation,
did your firm’s consumer demand for single-family \strong{refinance} mortgages
go up, go down, or stay the same?
\itemize{
\item Part a refers to GSE Eligible mortgages
\item Part b refers to Non-GSE Eligible mortgages
\item Part c refers to government (i.e. Federal Housing Administration
or Department of Veteran Affairs) mortgages.
}
\item 14 -- Over the \strong{next three months}, apart from normal seasonal variation,
do you expect your firm’s consumer demand for single-family
\strong{purchase} mortgages to go up, go down, or stay the same?
\itemize{
\item Part a refers to GSE Eligible mortgages
\item Part b refers to Non-GSE Eligible mortgages
\item Part c refers to government (i.e. Federal Housing Administration
or Department of Veteran Affairs) mortgages.
}
\item 18 -- Over the \strong{next three months}, apart from normal seasonal variation,
do you expect your firm’s consumer demand for single-family
\strong{refinance} mortgages to go up, go down, or stay the same?
\itemize{
\item Part a refers to GSE Eligible mortgages
\item Part b refers to Non-GSE Eligible mortgages
\item Part c refers to government (i.e. Federal Housing Administration
or Department of Veteran Affairs) mortgages.
}
\item 22 -- Over the \strong{next three months}, how much do you expect your firm's profit
margin to change for its single-family mortgage production?
\item 27 -- Over the \strong{past three months}, how did your firm’s credit standards for
approving consumer applications for mortgage loans change (across both
purchase mortgages and refinance mortgages)? Please answer for GSE Eligible
mortgages, Non-GSE Eligible mortgages, and Government mortgages.
\itemize{
\item Part a refers to GSE Eligible mortgages
\item Part b refers to Non-GSE Eligible mortgages
\item Part c refers to government (i.e. Federal Housing Administration
or Department of Veteran Affairs) mortgages.
\item Note that "up" refers to loosening of credit standards
and "down" refers to tightening of credit standards.
}
\item 31 -- Over the \strong{next three months}, how did your firm’s credit standards for
approving consumer applications for mortgage loans change (across both
purchase mortgages and refinance mortgages)? Please answer for GSE Eligible
mortgages, Non-GSE Eligible mortgages, and Government mortgages.
\itemize{
\item Part a refers to GSE Eligible mortgages
\item Part b refers to Non-GSE Eligible mortgages
\item Part c refers to government (i.e. Federal Housing Administration
or Department of Veteran Affairs) mortgages.
\item Note that "up" refers to loosening of credit standards
and "down" refers to tightening of credit standards.
}
}
}
\examples{
result <- get_lender_sentiment()
}
\seealso{
\link{lender_sentiment_questions}
}
