% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhs_data.R
\name{get_nhs_data}
\alias{get_nhs_data}
\title{Get National Housing Survey data}
\usage{
get_nhs_data()
}
\value{
A \code{tbl_df} containing the following columns:
\itemize{
\item \code{survey_date} -- The date of the survey
\item \code{question_id} -- The ID of the question asked
\item \code{question_ans} -- The answer associated with the question in \code{question_id}
\item \code{percent_ans} -- The proportion of respondents who answered \code{question_id}
as \code{question_ans}.
}
}
\description{
Get National Housing Survey data
}
\examples{
nhs_data <- get_nhs_data()
}
\seealso{
\link{nhs_questions}, \link{get_hpsi_data}
}
