% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Call a particular URL from Fannie Mae's public API.}
\usage{
get_url(url)
}
\arguments{
\item{url}{The URL. This should start with "https://api.theexchange.fanniemae.com/".}
}
\value{
A list with the following elements:
\itemize{
\item content -- The parsed JSON returned from API call.
Note that the object is returned as a nested list.
\item url -- The URL used to make the call
\item response -- The response object created by \code{httr::GET}
}
}
\description{
Call a particular URL from Fannie Mae's public API.
}
\examples{
get_url("https://api.theexchange.fanniemae.com/v1/mortgage-lender-sentiment/results")
}
