% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lender_sentiment_questions}
\alias{lender_sentiment_questions}
\title{Question look-up for Mortgage Lender Sentiment Survey data}
\format{A data frame with 19 rows and 5 variables: \code{question}, \code{question_num},
\code{question_sub}, \code{loan_type}, and \code{question_short}}
\usage{
lender_sentiment_questions
}
\description{
This is a look-up for the questions asked in Fannie Mae's mortgage
lender sentiment survey. The questionnaire can be found on
\href{http://www.fanniemae.com/resources/file/research/mlss/pdf/mortgage-lender-sentiment-survey-questionnaire-q42018.pdf}{Fannie Mae's website}.
}
\keyword{datasets}
