% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhs_questions}
\alias{nhs_questions}
\title{Question look-up for National Housing Survey data}
\format{A data frame with 22 rows and 3 columns: \code{question_id},
\code{question_desc}, and \code{question_ans}}
\usage{
nhs_questions
}
\description{
This is a look-up of the questions asked in Fannie Mae's
National Housing Survey (NHS). The questionnaire can be found on
\href{http://www.fanniemae.com/resources/file/research/housingsurvey/pdf/nhsquestionnaire-120517.pdf}{Fannie Mae's website}.
}
\keyword{datasets}
