% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the API authorization key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{The API key or user token as a character. This can be obtained by
from your account profile at \url{https://developer.theexchange.fanniemae.com/}.}
}
\value{
Invisibly returns the the R environment variable \code{fannieapi_key},
which is set as the \code{api_key} parameter in this function.
}
\description{
This function creates an R environment variable called \code{fannieapi_key},
which is used by subsequent functions to make API calls.
}
\examples{
set_api_key("abc123")
Sys.getenv("fannieapi_key")
}
