% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator.R
\name{get_indicator}
\alias{get_indicator}
\title{Get economic or housing indicators}
\usage{
get_indicator(category = c("economic", "housing"), indicator)
}
\arguments{
\item{category}{Either "economic" or "housing"}

\item{indicator}{One of the indicator series available under housing or economic indicators. See 'Details' for options.}
}
\value{
A \code{grouped_df} that is grouped by the column \code{report_date}.
}
\description{
Get economic or housing indicators
}
\details{
This function has access to the following indicator series for housing specific data:
\itemize{
\item gross-domestic-product
\item personal-consumption-expenditures
\item residential-fixed-investment
\item business-fixed-investment
\item government-consumption-and-investment
\item net-exports
\item change-in-business-inventories
\item gdp-price-index
\item consumer-price-index
\item consumer-price-index-excl-food-and-energy
\item personal-chain-expenditures-chain-price-index
\item personal-chain-expenditures-chain-price-index-excl-food-and-energy
\item unemployment-rate-percent
\item employment-total-nonfarm-percent-change-quarterly-saar-annual-q4-q4
\item federal-funds-rate
\item 1-year-treasury-note-yield
\item 10-year-treasury-bond-yield
}
}
\examples{
get_indicator("economic", "gross-domestic-product")
get_indicator("housing", "housing-starts-single-family")
}
\seealso{
\code{\link[=get_url]{get_url()}}
}
