% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{A call from Fannie Mae's public API given the API's path.}
\usage{
get_url(path)
}
\arguments{
\item{path}{This is the API's path after the origin URL.}
}
\value{
A list with the following elements:
\itemize{
\item content -- The parsed JSON returned from API call.
Note that the object is returned as a nested list.
\item url -- The URL used to make the call
\item response -- The response object created by \code{\link[httr:GET]{httr::GET()}}
}
}
\description{
A call from Fannie Mae's public API given the API's path.
}
\examples{
get_url("/v1/mortgage-lender-sentiment/results")
}
